Restructure Rust Backend to Support JBsDiff
JBsDiff structures the patch files differently. In classic C BsDiff, the
patch files are structured like this:

{| ctrl_block | delta_data | extra_data |} X N Blocks

In JBsDiff, the layout is

{| ctrl_block |} X N Blocks
{|delta_data|} X N Blocks
{|extra_data|} X N Blocks

In order to allow for this, the rust backend is abstracted into 3
Streams, one for each block. The protocol functions in the library root
use this to implement both. This means, however, that jbsdiff fromat is
not supported by the C backend.

The integration tests have changed as well. Now, instead of looking for
patch equivalence (which may not happen depending on compression
settings), there are effectively two tests.
1. Create a patch with the integration and apply it with rust
2. Create a patch with rust and apply it with the integration

If both of these produce the correct resulting data, integration is
probable, but no tests are perfect :-(

+++ test_setup.sh
@@ -3,11 +3,16 @@
 # Compile C bsdiff-43
 rm -rf target/c
 mkdir target/c
-cc bsdiff-43/bsdiff.c -DBSDIFF_EXECUTABLE -lbz2 -o target/c/bsdiff
-cc bsdiff-43/bspatch.c -DBSPATCH_EXECUTABLE -lbz2 -o target/c/bspatch
+cc bsdiff-43/bsdiff.c -DBSDIFF_EXECUTABLE -lbz2 -o target/c/bsdiff && \
+    cc bsdiff-43/bspatch.c -DBSPATCH_EXECUTABLE -lbz2 -o target/c/bspatch || echo "BsDiff C Build Failed!"
+
 
 # Compile Java jbsdiff
-mvn -f jbsdiff/pom.xml clean package
-rm -rf target/java
-mkdir target/java
-ln jbsdiff/target/jbsdiff-*.jar target/java/jbsdiff
+if mvn -f jbsdiff/pom.xml clean package
+then
+    rm -rf target/java
+    mkdir target/java
+    ln jbsdiff/target/jbsdiff-*.jar target/java/jbsdiff
+else
+    echo "BsDiff Java Build Failed!"
+fi
